Change linker to reuse code space.
