+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
     volumes:
       - ./server:/bluescript/server
       - /bluescript/server/node_modules
-      # - /bluescript/server/temp-files
+      - /bluescript/server/temp-files
       - ./modules:/bluescript/modules
       - ./microcontroller:/bluescript/microcontroller
     working_dir: /bluescript/server
