+++ server/src/compiler/compiler.ts
@@ -1,55 +1,16 @@
-import {MemoryBlock, MemoryRegion, ReusableMemoryRegion} from "./memory-region";
+import {MemoryBlock, ShadowMemory} from "./shadow-memory";
 import {ElfReader, SECTION_TYPE, Symbol} from "./elf-reader";
 import * as fs from "fs";
 import {execSync} from "child_process";
 import {
   LinerScriptMemoryAttribute,
-  LinkerScript3,
+  LinkerScript,
   LinkerScriptMemoryRegion,
   LinkerScriptSection
-} from "./linker-script3";
+} from "./linker-script";
 
-
-export type MemoryAddresses = {
-  iram:{address:number, size:number},
-  dram:{address:number, size:number},
-  flash:{address:number, size:number},
-}
-
-type MemoryUpdate = {
-  blocks: {address: number, data: string, isFlash: boolean}[],
-  entryPoints: number[]
-}
 
 const EXTERNAL_SYMBOL_SECTION = '.external_symbols'
-
-export class ShadowMemory {
-  public readonly iram: ReusableMemoryRegion;
-  public readonly dram: MemoryRegion;
-  public readonly flash: MemoryRegion;
-  public readonly symbols:Map<string, Symbol> = new Map<string, Symbol>();
-  public readonly componentsInfo = new ESPIDFComponentsInfo();
-  private updates: MemoryUpdate = {blocks: [], entryPoints: []};
-
-  constructor(bsRuntimePath: string, addresses: MemoryAddresses) {
-    const bsRuntime = new ElfReader(bsRuntimePath);
-    bsRuntime.readAllSymbols().forEach(symbol => {this.symbols.set(symbol.name, symbol)});
-    this.iram = new ReusableMemoryRegion('IRAM', addresses.iram.address, addresses.iram.size);
-    this.dram = new MemoryRegion('DRAM', addresses.dram.address, addresses.dram.size);
-    this.flash = new MemoryRegion('Flash', addresses.flash.address, addresses.flash.size);
-  }
-
-  setUpdate(blocks: {address: number, data: string, isFlash: boolean}[], entryPoint: number) {
-    this.updates.blocks = this.updates.blocks.concat(blocks);
-    this.updates.entryPoints.push(entryPoint)
-  }
-
-  getUpdates() {
-    const updates = this.updates;
-    this.updates = {blocks: [], entryPoints: []};
-    return updates;
-  }
-}
 
 
 export class Compiler {
@@ -57,6 +18,9 @@ export class Compiler {
   private readonly OBJ_FILE = './temp-files/code.o';
   private readonly LINKER_SCRIPT = './temp-files/linkerscript.ld';
   private readonly LINKED_ELF = './temp-files/code';
+  private readonly DATA_SECTIONS = ['.data', '.data.*', '.rodata', '.rodata.*', '.bss', '.bss.*', '.dram*'];
+  private readonly DRAM_SECTION_NAME = '.dram';
+
   constructor(private compilerPath = '') {}
 
   compile(shadowMemory: ShadowMemory, src: string, entryPointName: string) {
@@ -67,21 +31,22 @@ export class Compiler {
 
   private _compile(src: string) {
     fs.writeFileSync(this.C_FILE, src);
-    execSync(`${this.compilerPath}xtensa-esp32-elf-gcc -c -O2 ${this.C_FILE} -o ${this.OBJ_FILE} -w -fno-common -ffunction-sections -mtext-section-literals -mlongcalls`);
+    execSync(`${this.compilerPath}xtensa-esp32-elf-gcc -c -O2 ${this.C_FILE} -o ${this.OBJ_FILE} -w -fno-common -ffunction-sections -mtext-section-literals -mlongcalls -fno-zero-initialized-in-bss`);
     return new ElfReader(this.OBJ_FILE);
   }
 
   private link(shadowMemory: ShadowMemory, objFile: ElfReader, entryPointName: string) {
     this.freeIram(shadowMemory, objFile);
-    const memoryBlocks = this.allocate(shadowMemory, objFile);
+    const iramMemoryBlocks = this.allocateIram(shadowMemory, objFile);
     const externalSymbols = this.readExternalSymbols(shadowMemory, objFile);
-    const linkerScript = this.generateLinkerScript(memoryBlocks.iram, memoryBlocks.dram, externalSymbols, objFile, entryPointName);
+    const linkerScript = this.generateLinkerScript(shadowMemory, iramMemoryBlocks, externalSymbols, objFile, entryPointName);
     const linkedElf = this._link(linkerScript);
     this.freeEntryPoint(shadowMemory, entryPointName);
+    this.allocateDram(shadowMemory, linkedElf);
     return linkedElf;
   }
 
-  private _link(linkerScript: LinkerScript3) {
+  private _link(linkerScript: LinkerScript) {
     linkerScript.save(this.LINKER_SCRIPT)
     execSync(`${this.compilerPath}xtensa-esp32-elf-ld -o ${this.LINKED_ELF} -T ${this.LINKER_SCRIPT}`);
     return new ElfReader(this.LINKED_ELF);
@@ -112,7 +77,7 @@ export class Compiler {
   }
 
   private generateLinkerScript(
-    iramBlocks: MemoryBlock[], dramBlocks: MemoryBlock[],
+    shadowMemory: ShadowMemory, iramBlocks: MemoryBlock[],
     externalSymbols: Symbol[], objFile: ElfReader, entryPointName: string) {
     const memories: LinkerScriptMemoryRegion[] = [];
     const sections: LinkerScriptSection[] = [];
@@ -134,36 +99,36 @@ export class Compiler {
         [new LinerScriptMemoryAttribute('executable')],
         block.address, block.size);
       const section = new LinkerScriptSection(`.iram${index}`, memoryRegion)
-        .section(objFile.filePath, [block.sectionName ?? '']);
+        .section(objFile.filePath, [block.sectionName ?? ''])
+        .align(4);
       memories.push(memoryRegion);
       sections.push(section);
     });
-    dramBlocks.forEach((block , index) => {
-      const memoryRegion = new LinkerScriptMemoryRegion(
-        `DRAM${index}`,
-        [new LinerScriptMemoryAttribute('read/write'), new LinerScriptMemoryAttribute('readonly')],
-        block.address, block.size);
-      const section = new LinkerScriptSection(`.dram${index}`, memoryRegion)
-        .section(objFile.filePath, [block.sectionName ?? '']);
-      memories.push(memoryRegion);
-      sections.push(section);
-    });
+    const dramMemory = new LinkerScriptMemoryRegion(
+      'DRAM',
+      [new LinerScriptMemoryAttribute('read/write'), new LinerScriptMemoryAttribute('readonly')],
+      shadowMemory.dram.getNextAddress(),
+      1000000);
+    memories.push(dramMemory);
+    sections.push(new LinkerScriptSection(this.DRAM_SECTION_NAME, dramMemory)
+      .section(objFile.filePath, this.DATA_SECTIONS))
 
-    return new LinkerScript3()
+    return new LinkerScript()
       .input([objFile.filePath])
       .entry(entryPointName)
       .memory(memories)
       .sections(sections)
   }
 
-  private allocate(shadowMemory: ShadowMemory, objFile: ElfReader) {
+  private allocateIram(shadowMemory: ShadowMemory, objFile: ElfReader) {
     const executableSections = objFile.readSections(SECTION_TYPE.EXECUTABLE).filter(section => section.size !== 0);
-    const dataSections = objFile.readSections(SECTION_TYPE.WRITABLE)
-      .concat(objFile.readSections(SECTION_TYPE.READONLY))
-      .filter(section => section.size !== 0);
-    const iramMemoryBlock = executableSections.map(section => shadowMemory.iram.allocate(section));
-    const dramMemoryBlock = dataSections.map(section => shadowMemory.dram.allocate(section));
-    return {iram: iramMemoryBlock, dram: dramMemoryBlock};
+    return executableSections.map(section => shadowMemory.iram.allocate(section));
+  }
+
+  private allocateDram(shadowMemory: ShadowMemory, elf: ElfReader) {
+    const dramSection = elf.readSectionByName(this.DRAM_SECTION_NAME);
+    if (dramSection !== undefined)
+      shadowMemory.dram.allocate(dramSection);
   }
 
   private freeEntryPoint(shadowMemory: ShadowMemory, entryPointName: string) {
@@ -174,12 +139,12 @@ export class Compiler {
     const funcSymbols = objFile.readFunctions();
     funcSymbols.forEach(symbol => {
       if (shadowMemory.symbols.has(symbol.name))
-        shadowMemory.iram.free(this.funcNameToSectionName(symbol.name))
+        shadowMemory.iram.free(this.funcNameToSectionName(symbol.name));
     })
   }
 
   private funcNameToSectionName(funcName: string) {
-    return `.text.${funcName}`
+    return `.text.${funcName}`;
   }
 }
 
@@ -206,7 +171,7 @@ export class ModuleCompiler {
     return linkedElf;
   }
 
-  private _link(linkerScript: LinkerScript3) {
+  private _link(linkerScript: LinkerScript) {
     linkerScript.save(this.LINKER_SCRIPT);
     execSync(`${this.compilerPath}xtensa-esp32-elf-ld -o ${this.LINKED_ELF} -T ${this.LINKER_SCRIPT} --gc-sections`);
     return new ElfReader(this.LINKED_ELF);
@@ -268,49 +233,10 @@ export class ModuleCompiler {
         .section('*', this.EXECUTABLE_SECTIONS),
       externalSymbolSection
     ];
-    return new LinkerScript3()
+    return new LinkerScript()
       .input(shadowMemory.componentsInfo.getComponentsPath(moduleName))
       .memory(memories)
       .sections(sections)
       .entry(entryPointName)
   }
 }
-
-
-class ESPIDFComponentsInfo {
-  static DEPENDENCIES_FILE_PATH = '../microcontroller/ports/esp32/build/project_description.json'
-  static COMPONENTS_PATH_PREFIX = /^.*microcontroller\/ports\/esp32\/build/
-  static RELATIVE_PATH_COMMON = '../microcontroller/ports/esp32/build'
-
-  private dependenciesInfo: {[key: string]: {file: string, reqs: string[], priv_reqs: string[]}}
-
-  constructor() {
-    this.dependenciesInfo = JSON.parse(fs.readFileSync(ESPIDFComponentsInfo.DEPENDENCIES_FILE_PATH).toString()).build_component_info
-  }
-
-  public getComponentsPath(rootComponent: string): string[] {
-    let tmp = [rootComponent]
-    let visited = new Set<string>()
-    const componentPaths:string[] = []
-    while(tmp.length > 0) {
-      let curr = tmp.shift() as string
-      visited.add(curr)
-      tmp = tmp.concat(
-        this.dependenciesInfo[curr].priv_reqs.filter((r:string) => !visited.has(r)),
-        this.dependenciesInfo[curr].reqs.filter((r:string) => !visited.has(r))
-      )
-      if (this.dependenciesInfo[curr].file !== undefined && this.dependenciesInfo[curr].file !== '')
-        componentPaths.push(this.convertAbsoluteToRelative(this.dependenciesInfo[curr].file))
-    }
-    return componentPaths
-  }
-
-  public getComponentPath(componentName: string) {
-    return this.convertAbsoluteToRelative(this.dependenciesInfo[componentName].file)
-  }
-
-  private convertAbsoluteToRelative(absolutePath: string) {
-    return absolutePath.replace(ESPIDFComponentsInfo.COMPONENTS_PATH_PREFIX, ESPIDFComponentsInfo.RELATIVE_PATH_COMMON)
-  }
-}
-
